Create LICENSE
I found this repository has no licence, so it's hard for people to use the learnings from it and apply them. I used choosealicense.com to propose a simple, permissive one. Feel free to choose any other.  Issues are not open, so making a PR was the only way to discuss this topic.

Cheers!